separate main functions from governance scripts
