+++ script/hardhat/governance/deploy-governor-bravo-delegate.js
@@ -11,8 +11,4 @@ const main = async() => {
     console.log(`deployer: ${deployer} deployed GovernorBravoDelegate at address: ${governorBravoDelegateContractInstance.address}`);
 };
 
-main().then(() => process.exit(0))
-.catch((error) => {
-  console.error(error);
-  process.exit(1);
-});
+module.exports = main;
