+++ script/hardhat/governance/run/verify-governor-bravo-delegate.js
@@ -1,0 +1,7 @@
+const main = require('../verify-governor-bravo-delegate');
+
+main().then(() => process.exit(0))
+  .catch((error) => {
+    console.error(error);
+    process.exit(1);
+  });
