+++ script/hardhat/governance/verify-governor-bravo-delegate.js
@@ -1,5 +1,5 @@
 require("dotenv").config();
-const contractConfigData = require("../../networks/testnet.json");
+const contractConfigData = require("../../../networks/testnet.json");
 const hre = require("hardhat");
 
 const main = async () => {
@@ -10,8 +10,4 @@ const main = async () => {
   });
 };
 
-main().then(() => process.exit(0))
-  .catch((error) => {
-    console.error(error);
-    process.exit(1);
-  });
+module.exports = main;
